Ajustando relacionamento fornecedor, impressora e consumiveis

+++ database/sql/CREATE/002_CREATE_tipo_fornecedor.sql
@@ -1,7 +1,7 @@
 IF NOT EXISTS(SELECT 1 FROM sys.tables WHERE NAME = 'tipo_fornecedor')
 BEGIN
   CREATE TABLE alugel_impressora.dbo.tipo_fornecedor (
-    id_tipo_fornecedor BIGINT NOT NULL IDENTITY(1,1),
+    id_tipo_fornecedor INT NOT NULL IDENTITY(1,1),
     cd_tipo_fornecedor INT NOT NULL UNIQUE,
     nm_tipo_fornecedor VARCHAR(30) NOT NULL,
     dt_inclusao DATETIME NOT NULL DEFAULT GETDATE(),
